requirements.txt
discord
streamrip
logging
validators
python-dotenv